Add initial support for bhyve hypervisor to the nova libvirt driver
This is very initial, basic support for FreeBSD and bhyve hypervisor. It is
intended to provide a codebase for further development. It requires a
development version of the libvirt library from the official git repository.
Due to very early support for bhyve in libvirt, it comes with many
limitations that includes:
- Only FreeBSD guest OSes can be used
- No support for the config drive
- Only one disk and one Ethernet interface
- No pause/resume functionality
- No VM migration support
- No files injection ito VM's filestystem
- Only works with nova-network (that uses freebsd_net driver) under
  Flat/FlatDHCP multi-host mode